Many changes
Updated screen size selection in hopes that this will scale correctly on mobile devices.  Added a few bits into stylesheet (regarding size), changed some things in setupnumbers.js, setupgrid.js, index.html and gameplay.js.  Removed .container class because it messed the mobile viewport up, because of this had to add in a fudge of sorts to get the SVG circles to display (approximately) in the middle of the screen, in landscape or portrait.

TO DO: Add images instead of white background, Start screen, end screen, high scores? timer? grid size select.